requirements.txt
bcrypt==3.1.6
circus==0.15.0
cronex==0.1.3.1
Flask==1.0.2
Flask-Admin==1.5.3
flask-restplus==0.12.1
flask-security==3.0.0
Flask-SQLAlchemy==2.4.0
freezegun==0.3.11
kafka-python==1.4.6
msgpack==0.6.1
python-dateutil==2.8.0
python-dotenv==0.10.1
redis==3.2.1
requests_flask_adapter==0.0.1
smart-getenv==1.1.0
SQLAlchemy==1.3.3
tenacity==5.0.4